[#8] chore : chattingRoomPage 폴더 이름 변경, chattingRoom component로 변경
