+++ catch-study/src/App.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import ChattingPage from './pages/ChattingPage/ChattingPage';
-import ChattingRoomPage from './pages/ChattingRoom/ChattingRoomPage';
+import ChattingRoomPage from './pages/ChattingRoomPage/ChattingRoomPage';
 import MainPage from './pages/MainPage/MainPage';
 
 const App: React.FC = () => {
