+++ catch-study/src/pages/ChattingRoomPage/ChattingRoomPage.tsx
@@ -1,0 +1,13 @@
+import ChattingRoom from '../../components/ChattingRoom/ChattingRoom';
+import Topbar from '../../components/Topbar/Topbar';
+
+const ChattingRoomPage: React.FC = () => {
+  return (
+    <>
+      <Topbar />
+      <ChattingRoom />
+    </>
+  );
+};
+
+export default ChattingRoomPage;
